Removed proto.Any dependency in scheduler.Schedule
Previously we relied on protobuf's Any to store  the actions decided upon by the scheduler. Due to the implementation of Protobufs implementation this required us to constantly marshal and unmarshal these actions into bytes. By reverting back to a typed data structure containing fields to set each field we remove the need for this.